requirements.txt
streamlit==1.55.0
pandas==2.3.3
duckdb==1.5.0
folium==0.20.0
streamlit-folium==0.26.2
requests==2.32.5
python-dotenv==1.2.2
anthropic==0.50.0
